+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.8.0] - 2019-06-15
+
 ### Added
 
 * Add facebook share preview for content editors when editing Opengraph meta data
@@ -91,7 +93,8 @@ allow installations via Composer
 
 * Initial release of the module 🐣
 
-[Unreleased]: https://github.com/wanze/SeoMaestro/compare/v0.7.0...HEAD
+[Unreleased]: https://github.com/wanze/SeoMaestro/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/wanze/SeoMaestro/releases/tag/v0.8.0
 [0.7.0]: https://github.com/wanze/SeoMaestro/releases/tag/v0.7.0
 [0.6.0]: https://github.com/wanze/SeoMaestro/releases/tag/v0.6.0
 [0.5.0]: https://github.com/wanze/SeoMaestro/releases/tag/v0.5.0